refactor(desktop): extract pure host-service helpers to host-service-utils
